refactor(nuxt,vite): move externals/deep-imports vite plugin -> vite builder
