+++ packages/nuxt/meta.d.mts
