+++ packages/nuxt/meta.mjs
