+++ tsconfig.json
@@ -43,7 +43,7 @@
     "**/dist/**",
     "**/.nuxt/**",
     "packages/nuxt/src/app/types/**",
-    "packages/nuxt/src/meta.d.mts",
+    "packages/nuxt/meta.d.mts",
     "**/nuxt.d.ts",
     "**/examples/**",
     "**/docs/**",
